+++ l10n_br_nfse/models/document.py
@@ -21,8 +21,6 @@ from ..constants.nfse import (
     OPERATION_NATURE,
     RPS_TYPE,
     TAXATION_SPECIAL_REGIME,
-    RECEPCIONAR_LOTE_RPS,
-    CONSULTAR_SITUACAO_LOTE_RPS,
 )
 
 from .res_company import PROCESSADOR
@@ -178,95 +176,3 @@ class Document(models.Model):
             'intermediario_servico': None,
             'construcao_civil': None,
         }
-
-    @api.multi
-    def _eletronic_document_send(self):
-        super(Document, self)._eletronic_document_send()
-        for record in self.filtered(fiter_processador_edoc_nfse):
-            processador = record._processador_erpbrasil_nfse()
-
-            protocolo = record.protocolo_autorizacao
-            vals = dict()
-
-            if not protocolo:
-                for edoc in record.serialize():
-                    processo = None
-                    for p in processador.processar_documento(edoc):
-                        processo = p
-
-                        if processo.webservice in RECEPCIONAR_LOTE_RPS:
-                            if processo.resposta.Protocolo is None:
-                                mensagem_completa = ''
-                                if processo.resposta.ListaMensagemRetorno:
-                                    lista_msgs = \
-                                        processo.resposta.ListaMensagemRetorno
-                                    for mr in lista_msgs.MensagemRetorno:
-
-                                        correcao = ''
-                                        if mr.Correcao:
-                                            correcao = mr.Correcao
-
-                                        mensagem_completa += (
-                                            mr.Codigo + ' - ' +
-                                            mr.Mensagem +
-                                            ' - Correção: ' +
-                                            correcao + '\n'
-                                        )
-                                vals['edoc_error_message'] = mensagem_completa
-                                record.write(vals)
-                                return
-                            protocolo = processo.resposta.Protocolo
-
-                    if processo.webservice in CONSULTAR_SITUACAO_LOTE_RPS:
-                        vals['codigo_situacao'] = processo.resposta.Situacao
-            else:
-                vals['codigo_situacao'] = 4
-
-            if vals.get('codigo_situacao') == 1:
-                vals['motivo_situacao'] = _('Não Recebido')
-
-            elif vals.get('codigo_situacao') == 2:
-                vals['motivo_situacao'] = _('Lote ainda não processado')
-
-            elif vals.get('codigo_situacao') == 3:
-                vals['motivo_situacao'] = _('Procesado com Erro')
-
-            elif vals.get('codigo_situacao') == 4:
-                vals['motivo_situacao'] = _('Procesado com Sucesso')
-                vals['protocolo_autorizacao'] = protocolo
-
-            if vals.get('codigo_situacao') in (3, 4):
-                processo = processador.consultar_lote_rps(protocolo)
-
-                if processo.resposta:
-                    mensagem_completa = ''
-                    if processo.resposta.ListaMensagemRetorno:
-                        lista_msgs = processo.resposta.ListaMensagemRetorno
-                        for mr in lista_msgs.MensagemRetorno:
-
-                            correcao = ''
-                            if mr.Correcao:
-                                correcao = mr.Correcao
-
-                            mensagem_completa += (
-                                mr.Codigo + ' - ' +
-                                mr.Mensagem +
-                                ' - Correção: ' +
-                                correcao + '\n'
-                            )
-                    vals['edoc_error_message'] = mensagem_completa
-
-                if processo.resposta.ListaNfse:
-                    xml_file = processador._generateds_to_string_etree(
-                        processo.resposta)[0]
-                    record.autorizacao_event_id.set_done(xml_file)
-                    for comp in processo.resposta.ListaNfse.CompNfse:
-                        vals['number'] = comp.Nfse.InfNfse.Numero
-                        vals['data_hora_autorizacao'] = \
-                            comp.Nfse.InfNfse.DataEmissao
-                        vals['verify_code'] = \
-                            comp.Nfse.InfNfse.CodigoVerificacao
-                    record._change_state(SITUACAO_EDOC_AUTORIZADA)
-
-            record.write(vals)
-        return
